task: moving cli into tools
