--- cli/db/go.mod
@@ -1,10 +1,0 @@
-module github.com/leapkit/leapkit/cli/db
-
-go 1.22
-
-require (
-	github.com/leapkit/leapkit/core v0.0.43
-	github.com/lib/pq v1.10.9
-	github.com/mattn/go-sqlite3 v1.14.23
-	github.com/spf13/pflag v1.0.6
-)
