--- cli/dev/go.mod
@@ -1,10 +1,0 @@
-module github.com/leapkit/leapkit/cli/dev
-
-go 1.22
-
-require (
-	github.com/fsnotify/fsnotify v1.8.0
-	github.com/spf13/pflag v1.0.6
-)
-
-require golang.org/x/sys v0.21.0 // indirect
