--- cli/new/go.mod
@@ -1,3 +1,0 @@
-module github.com/leapkit/leapkit/cli/new
-
-go 1.22
