+++ template/go.mod
@@ -9,7 +9,7 @@ require (
 
 require (
 	github.com/gobuffalo/flect v1.0.2 // indirect
-	github.com/gobuffalo/plush/v5 v5.0.1 // indirect
+	github.com/gobuffalo/plush/v5 v5.0.2 // indirect
 	github.com/gorilla/securecookie v1.1.2 // indirect
 	github.com/gorilla/sessions v1.3.0 // indirect
 )
