+++ tools/db/internal/database/migration.sql.tmpl
