+++ tools/dev/internal/rebuilder/writer.go
