+++ tools/go.mod
@@ -1,0 +1,15 @@
+module github.com/leapkit/leapkit/tools
+
+go 1.22
+
+require (
+	github.com/fsnotify/fsnotify v1.8.0
+	github.com/leapkit/leapkit/cli/db v0.0.0-20250221133609-cb80315527d1
+	github.com/leapkit/leapkit/cli/dev v0.0.0-20250221133609-cb80315527d1
+	github.com/leapkit/leapkit/core v0.0.43
+	github.com/lib/pq v1.10.9
+	github.com/mattn/go-sqlite3 v1.14.24
+	github.com/spf13/pflag v1.0.6
+)
+
+require golang.org/x/sys v0.21.0 // indirect
