+++ tools/new/main.go
